// src/api/generated.ts
export interface TodoResponse {
  id: string;
  title: string;
  userId: string;
  completed: boolean;
  createdAt: string;
  description?: string;
}

export interface User {
  username: string;
  password: string;
}

export interface UserResponse {
  id: string;
  username: string;
  createdAt: string;
}

export type RefreshTokenBody = {
  refreshToken: string;
};

export type RefreshToken200 = {
  /** The newly generated access token. */
  accessToken?: string;
};

export type GetTodoList200 = {
  todos: TodoResponse[];
};